refactor(connectors): rename google_keep/google_tasks → googlekeep/googletasks
Underscored package names violate Go's package-naming convention and
were the root cause of CodeFactor's `package-naming` flag we'd been
trying to suppress via .codefactor.yml. The proper fix per
`feedback_no_lazy_lint_suppressions` is to rename the packages, not
expand suppressions.

Mechanical changes:
- `internal/connectors/google_keep/` → `internal/connectors/googlekeep/`
- `internal/connectors/google_tasks/` → `internal/connectors/googletasks/`
- `package google_keep` → `package googlekeep` (and `..._test` external test packages)
- `package google_tasks` → `package googletasks` (and `..._test`)
- All import paths swept; the existing alias names (`googlekeep`, `googletasks`)
  are now redundant and were stripped — Go infers the package name from the path.
- `//revive:disable:var-naming` directives in the four affected files removed:
  the rule no longer trips because the names are conformant.
- `.semgrep/rules.yml` path-exclusions updated to the new directory names so
  pre-existing in-package rule waivers (Keep gateway error-discard idioms,
  etc.) still apply.

Reverts the .codefactor.yml suppressions added earlier in this session
(connector-package paths in both top-level `exclude_patterns` and
per-tool `golint`/`revive` excluded_paths). Restores CodeFactor coverage
on those packages: complexity, dead-code, security, and other rules
now run normally — only the now-moot `package-naming` rule needed
suppression, and that's gone with the rename.

Frontmatter persistence keys (`wiki.connectors.google_keep.*`,
`wiki.connectors.google_tasks.*`) are intentionally NOT renamed: those
mirror the `ConnectorKindGoogleKeep` / `ConnectorKindGoogleTasks` proto
enum slugs used as on-disk identifiers. Renaming them would require a
migration on every operator's profile page. Same reasoning for error-
message prefixes (`"google_tasks: ..."`).
