+++ .semgrep/rules.yml
@@ -109,9 +109,9 @@ rules:
         - vendor/
         - "*_test.go"
         # Parallel error-classification work in flight; re-include after merge.
-        - internal/connectors/google_tasks/gateway/
-        - internal/connectors/google_tasks/sync/
-        - internal/connectors/google_keep/gateway/
+        - internal/connectors/googletasks/gateway/
+        - internal/connectors/googletasks/sync/
+        - internal/connectors/googlekeep/gateway/
 
   - id: go.error-classified-by-status-code-only
     languages: [go]
@@ -148,9 +148,9 @@ rules:
         - "*_test.go"
         - cmd/
         # Parallel error-classification work in flight; re-include after merge.
-        - internal/connectors/google_tasks/gateway/
-        - internal/connectors/google_tasks/sync/
-        - internal/connectors/google_keep/gateway/
+        - internal/connectors/googletasks/gateway/
+        - internal/connectors/googletasks/sync/
+        - internal/connectors/googlekeep/gateway/
 
   - id: go.handler-returns-nil-after-error-logged
     languages: [go]
@@ -191,9 +191,9 @@ rules:
         - "*_test.go"
         - main.go
         # Parallel error-classification work in flight; re-include after merge.
-        - internal/connectors/google_tasks/gateway/
-        - internal/connectors/google_tasks/sync/
-        - internal/connectors/google_keep/gateway/
+        - internal/connectors/googletasks/gateway/
+        - internal/connectors/googletasks/sync/
+        - internal/connectors/googlekeep/gateway/
 
   - id: go.errors-As-without-target-pointer
     languages: [go]
@@ -247,9 +247,9 @@ rules:
         # actively reworked by parallel error-classification work. They will be
         # re-included once that work merges and existing discards have been
         # given justifications inline.
-        - internal/connectors/google_tasks/gateway/
-        - internal/connectors/google_tasks/sync/
-        - internal/connectors/google_keep/gateway/
+        - internal/connectors/googletasks/gateway/
+        - internal/connectors/googletasks/sync/
+        - internal/connectors/googlekeep/gateway/
 
   # ============================================================
   # TypeScript / JavaScript Rules
