+++ internal/bootstrap/server.go
@@ -19,8 +19,8 @@ import (
 	"github.com/brendanjerwin/simple_wiki/internal/caldav"
 	"github.com/brendanjerwin/simple_wiki/internal/connectors"
 	"github.com/brendanjerwin/simple_wiki/internal/connectors/engine"
-	googletasks "github.com/brendanjerwin/simple_wiki/internal/connectors/google_tasks"
-	tasksgateway "github.com/brendanjerwin/simple_wiki/internal/connectors/google_tasks/gateway"
+	"github.com/brendanjerwin/simple_wiki/internal/connectors/googletasks"
+	tasksgateway "github.com/brendanjerwin/simple_wiki/internal/connectors/googletasks/gateway"
 	grpcapi "github.com/brendanjerwin/simple_wiki/internal/grpc/api/v1"
 	wikimcp "github.com/brendanjerwin/simple_wiki/internal/mcp"
 	"github.com/brendanjerwin/simple_wiki/internal/observability"
