+++ internal/connectors/googlekeep/credentials_test.go
@@ -1,5 +1,5 @@
 //revive:disable:dot-imports
-package google_keep_test
+package googlekeep_test
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	googlekeep "github.com/brendanjerwin/simple_wiki/internal/connectors/google_keep"
+	"github.com/brendanjerwin/simple_wiki/internal/connectors/googlekeep"
 	"github.com/brendanjerwin/simple_wiki/wikipage"
 )
