+++ internal/connectors/googlekeep/gateway/types.go
