+++ internal/connectors/googlekeep/translator/observed.go
