+++ internal/grpc/api/v1/connector_service.go
@@ -13,10 +13,10 @@ import (
 	apiv1 "github.com/brendanjerwin/simple_wiki/gen/go/api/v1"
 	"github.com/brendanjerwin/simple_wiki/internal/connectors"
 	"github.com/brendanjerwin/simple_wiki/internal/connectors/engine"
-	googlekeep "github.com/brendanjerwin/simple_wiki/internal/connectors/google_keep"
-	"github.com/brendanjerwin/simple_wiki/internal/connectors/google_keep/gateway"
-	googletasks "github.com/brendanjerwin/simple_wiki/internal/connectors/google_tasks"
-	tasksgateway "github.com/brendanjerwin/simple_wiki/internal/connectors/google_tasks/gateway"
+	"github.com/brendanjerwin/simple_wiki/internal/connectors/googlekeep"
+	"github.com/brendanjerwin/simple_wiki/internal/connectors/googlekeep/gateway"
+	"github.com/brendanjerwin/simple_wiki/internal/connectors/googletasks"
+	tasksgateway "github.com/brendanjerwin/simple_wiki/internal/connectors/googletasks/gateway"
 	"github.com/brendanjerwin/simple_wiki/tailscale"
 	"github.com/brendanjerwin/simple_wiki/wikipage"
 )
